# packages/contracts/openapi/openapi.yaml
components:
  schemas:
    CreateIngestKeyBody:
      additionalProperties: false
      properties:
        name:
          maxLength: 64
          minLength: 1
          type: string
      required:
        - name
      type: object
    CreateIngestKeyResponse:
      additionalProperties: false
      properties:
        createdAt:
          format: date-time
          type: string
        id:
          minLength: 1
          type: string
        name:
          maxLength: 64
          minLength: 1
          type: string
        secret:
          minLength: 1
          type: string
      required:
        - id
        - name
        - createdAt
        - secret
      type: object
...
    ErrorResponse:
      additionalProperties: false
      properties:
        error:
          additionalProperties: false
          properties:
            code:
              minLength: 1
              type: string
            message:
              minLength: 1
              type: string
          required:
            - code
            - message
          type: object
      required:
        - error
      type: object
...
    IngestKeyListResponse:
      items:
        additionalProperties: false
        properties:
          createdAt:
            format: date-time
            type: string
          id:
            minLength: 1
            type: string
          lastUsedAt:
            format: date-time
            type: string
          name:
            maxLength: 64
            minLength: 1
            type: string
        required:
          - id
          - name
          - createdAt
        type: object
      type: array
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: Short-lived (15 minute) JWT access token from login or refresh.
      scheme: bearer
      type: http
    ingestKeyAuth:
      description: Ingest credentials `keyId.secret`; every request must also carry valid X-Notify-Timestamp and X-Notify-Signature HMAC headers.
      scheme: bearer
      type: http
info:
  description: Single protocol source for the opencode-notify gateway, the OpenCode plugin, and the Flutter clients. Generated from the shared JSON schemas in @notify/contracts; do not edit by hand.
  license:
    identifier: MIT
    name: MIT
  title: opencode-notify gateway
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/ingest-keys:
    get:
      description: List items never include the key secret.
      operationId: listIngestKeys
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestKeyListResponse"
          description: Ingest keys without secrets.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid access token.
      security:
        - bearerAuth: []
      summary: List ingest keys.
      tags:
        - ingest-keys
    post:
      description: The secret is returned exactly once in this response and is never recoverable afterwards; the server stores only SHA-256(secret).
      operationId: createIngestKey
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIngestKeyBody"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateIngestKeyResponse"
          description: Created ingest key including the one-time secret.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Validation failed.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid access token.
      security:
        - bearerAuth: []
      summary: Create an ingest key.
      tags:
        - ingest-keys
  /v1/ingest-keys/{id}:
    delete:
      operationId: revokeIngestKey
      parameters:
        - description: Ingest-key identifier.
          in: path
          name: id
          required: true
          schema:
            minLength: 1
            type: string
      responses:
        "204":
          description: Ingest key revoked.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid access token.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unknown ingest key for this user.
      security:
        - bearerAuth: []
      summary: Revoke an ingest key.
      tags:
        - ingest-keys
...